chore(user): rename user composable appropriately
